Added the main functionality - currency conversion

+++ model/currency.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Currency struct {
-	ID           int64     `gorm:"primaryKey"`
+	ID           int       `gorm:"primaryKey"`
 	AdjustInd    string    `gorm:"size:1 ; column:adjustInd"`
 	Begin        time.Time `gorm:"type:date; column:begin"`
 	Decimals     int       `gorm:"not null; column:decimals"`
@@ -14,7 +14,7 @@ type Currency struct {
 	IsoCode      string    `gorm:"size:3; column:isoCode"`
 	RoundingUnit float64   `gorm:"not null; column:roundingUnit"`
 	LastUpdate   time.Time `gorm:"default:current_timestamp; column:last_update"`
-	CurrencyID   int64     `gorm:"not null; column:currencyId"`
+	CurrencyID   int       `gorm:"not null; column:currencyId"`
 }
 
 func (Currency) TableName() string {
